+++ rest/src/main/java/biz/karms/sinkit/rest/SinkitService.java
@@ -1,18 +1,13 @@
 package biz.karms.sinkit.rest;
 
-import biz.karms.sinkit.ejb.BlacklistedRecord;
-import biz.karms.sinkit.ejb.Rule;
-import biz.karms.sinkit.ejb.ServiceEJB;
-import biz.karms.sinkit.ejb.ArchiveServiceEJB;
-import biz.karms.sinkit.ejb.CoreServiceEJB;
+import biz.karms.sinkit.ejb.*;
 import biz.karms.sinkit.ioc.IoCRecord;
-
 import com.google.gson.GsonBuilder;
 
+import javax.ejb.DependsOn;
 import javax.ejb.EJB;
 import javax.enterprise.context.SessionScoped;
 import javax.inject.Inject;
-import java.io.IOException;
 import java.io.Serializable;
 import java.util.logging.Level;
 import java.util.logging.Logger;
@@ -24,31 +19,34 @@ import java.util.logging.Logger;
  */
 @SessionScoped
 public class SinkitService implements Serializable {
-    private static final long serialVersionUID = -940606926425474624L;
+    private static final long serialVersionUID = -940606926425473624L;
     public static final String ERR_MSG = "Error, please, check your input.";
 
     @EJB
-    private ServiceEJB serviceEJB;
+    private WebApiEJB webapiEJB;
 
     @EJB
     private CoreServiceEJB coreService;
+
+    @EJB
+    private DNSApiEJB dnsApiEJB;
 
     @Inject
     private Logger log;
 
     String createHelloMessage(final String name) {
-        return new GsonBuilder().create().toJson(serviceEJB.sayHello(name));
+        return new GsonBuilder().create().toJson(webapiEJB.sayHello(name));
     }
 
     String getStats() {
-        return new GsonBuilder().create().toJson(serviceEJB.getStats());
+        return new GsonBuilder().create().toJson(webapiEJB.getStats());
     }
 
     String putBlacklistedRecord(final String json) {
         try {
             log.log(Level.FINEST, "Received JSON [" + json + "]");
             BlacklistedRecord blacklistedRecord = new GsonBuilder().create().fromJson(json, BlacklistedRecord.class);
-            blacklistedRecord = serviceEJB.putBlacklistedRecord(blacklistedRecord);
+            blacklistedRecord = webapiEJB.putBlacklistedRecord(blacklistedRecord);
             if (blacklistedRecord != null) {
                 return new GsonBuilder().create().toJson(blacklistedRecord);
             } else {
@@ -61,15 +59,19 @@ public class SinkitService implements Serializable {
     }
 
     String getBlacklistedRecord(final String key) {
-        return new GsonBuilder().create().toJson(serviceEJB.getBlacklistedRecord(key));
+        return new GsonBuilder().create().toJson(webapiEJB.getBlacklistedRecord(key));
+    }
+
+    String getSinkHole(final String client, final String key) {
+        return new GsonBuilder().create().toJson(dnsApiEJB.getSinkHole(client, key));
     }
 
     String getBlacklistedRecordKeys() {
-        return new GsonBuilder().create().toJson(serviceEJB.getBlacklistedRecordKeys());
+        return new GsonBuilder().create().toJson(webapiEJB.getBlacklistedRecordKeys());
     }
 
     String deleteBlacklistedRecord(final String key) {
-        String message = serviceEJB.deleteBlacklistedRecord(key);
+        String message = webapiEJB.deleteBlacklistedRecord(key);
         if (message == null) {
             return new GsonBuilder().create().toJson(ERR_MSG);
         }
@@ -83,7 +85,7 @@ public class SinkitService implements Serializable {
             if (rule == null) {
                 return new GsonBuilder().create().toJson(ERR_MSG);
             }
-            return new GsonBuilder().create().toJson(serviceEJB.putRule(rule));
+            return new GsonBuilder().create().toJson(webapiEJB.putRule(rule));
         } catch (Exception e) {
             log.log(Level.SEVERE, "putRule", e);
             return new GsonBuilder().create().toJson(ERR_MSG);
@@ -91,22 +93,22 @@ public class SinkitService implements Serializable {
     }
 
     String getRules(final String clientIPAddress) {
-        return new GsonBuilder().create().toJson(serviceEJB.getRules(clientIPAddress));
+        return new GsonBuilder().create().toJson(webapiEJB.getRules(clientIPAddress));
     }
 
     String getRuleKeys() {
-        return new GsonBuilder().create().toJson(serviceEJB.getRuleKeys());
+        return new GsonBuilder().create().toJson(webapiEJB.getRuleKeys());
     }
 
     String deleteRule(final String cidrAddress) {
-        String message = serviceEJB.deleteRule(cidrAddress);
+        String message = webapiEJB.deleteRule(cidrAddress);
         if (message == null) {
             return new GsonBuilder().create().toJson(ERR_MSG);
         }
         return new GsonBuilder().create().toJson(message);
     }
 
-    String processIoCRecord(String jsonIoCRecord)  {
+    String processIoCRecord(String jsonIoCRecord) {
 
         String response;
         try {
@@ -117,7 +119,7 @@ public class SinkitService implements Serializable {
             }
             if (ioc.getSource() == null || (
                     ioc.getSource().getFQDN() == null && ioc.getSource().getIp() == null && ioc.getSource().getUrl() == null
-                    )) {
+            )) {
                 throw new Exception("IoC can't have both IP and Domain set as null");
             }
             if (ioc.getClassification() == null || ioc.getClassification().getType() == null) {
@@ -132,7 +134,7 @@ public class SinkitService implements Serializable {
             response = new GsonBuilder().setDateFormat(IoCRecord.DATE_FORMAT).create().toJson(ioc);
         } catch (Exception e) {
             e.printStackTrace();
-            log.log(Level.SEVERE,"IoC: " + jsonIoCRecord);
+            log.log(Level.SEVERE, "IoC: " + jsonIoCRecord);
             response = new GsonBuilder().setDateFormat(IoCRecord.DATE_FORMAT).create().toJson(e.getMessage());
         }
         return response;
